separates parse_simple_formula to utils.nameparser.py
