+++ src/exogibbs/thermo/stoichiometry.py
@@ -21,11 +21,69 @@ import pandas as pd
 # Add near the top with other regexes
 
 # Matches one or more trailing parenthetical annotation groups, e.g. "(CNN)", "(g)", "(NCN)"
-_ELNUM = re.compile(r"([A-Z][a-z]?)(\d*)")
 _PAREN_ANNOT_TAIL = re.compile(r"(?:\([A-Za-z0-9+\-]*\))+$")
 # Add this near the top with the other imports/regexes
 _ELECTRON_BASE = re.compile(r"^[eE](\d*)$")  # matches 'e', 'e1', 'E2', etc.
 _CHARGE = re.compile(r"^([A-Za-z0-9*]+)([+-]\d*)$")  # unchanged
+
+from exogibbs.utils.nameparser import parse_simple_formula
+
+def build_formula_matrix_from_JANAF(
+    df_molname: pd.DataFrame,
+    species_col: str = "JANAF",
+    *,
+    element_order: Optional[Sequence[str]] = None,
+    sanitize: bool = True,
+    species_names: str = "raw",  # NEW: "raw" | "clean" | "both"
+) -> Union[
+    Tuple[np.ndarray, List[str], List[str]],
+    Tuple[np.ndarray, List[str], List[str], List[str]]
+]:
+    """
+    Build a stoichiometric matrix A (elements x species) from df_molname.
+
+    Args:
+        species_names : {"raw", "clean", "both"}, default "raw"
+        Which species labels to return:
+        - "raw": return original strings (e.g., "C1N2(CNN)")
+        - "clean": return sanitized/normalized species (e.g., "C1N2")
+        - "both": return both (species_raw, species_clean)
+
+    Returns:
+        If species_names in {"raw", "clean"}: A, elements, species
+        If species_names == "both": A, elements, species_raw, species_clean
+    """
+    raw_species = df_molname[species_col].astype(str).tolist()
+
+    # Clean names for parsing / normalization (drop leading markers & trailing annotations)
+    species_clean = [sanitize_formula(s) if sanitize else s for s in raw_species]
+
+    # Parse using the clean base (never use raw for parsing)
+    parsed_list = [parse_formula_with_charge(s) for s in species_clean]
+
+    all_elements = set().union(*parsed_list) if parsed_list else set()
+    elements = sorted(all_elements) if element_order is None else list(element_order)
+
+    A = np.zeros((len(elements), len(species_clean)), dtype=np.int64)
+    elem_index = {e: i for i, e in enumerate(elements)}
+
+    for j, d in enumerate(parsed_list):
+        for e, v in d.items():
+            if e not in elem_index:
+                raise ValueError(
+                    f"Unexpected element '{e}' in species '{species_clean[j]}'."
+                )
+            A[elem_index[e], j] = int(v)
+
+    # Choose which species names to return (labels only; matrix columns already built)
+    if species_names == "raw":
+        return A, elements, raw_species
+    elif species_names == "clean":
+        return A, elements, species_clean
+    elif species_names == "both":
+        return A, elements, raw_species, species_clean
+    else:
+        raise ValueError("species_names must be 'raw', 'clean', or 'both'.")
 
 
 def sanitize_formula(s: str) -> str:
@@ -101,98 +159,5 @@ def parse_formula_with_charge(formula: str) -> Dict[str, int]:
         element_count_dict["e-"] = element_count_dict.get("e-", 0) - charge
 
     return element_count_dict
-
-
-def parse_simple_formula(formula: str) -> Dict[str, int]:
-    """
-    Parse a simple chemical formula string into a dict of element counts.
-
-    Examples:
-        "CH4"   -> {"C": 1, "H": 4}
-        "C1O2"  -> {"C": 1, "O": 2}
-        "H2O"   -> {"H": 2, "O": 1}
-
-    Notes:
-        - Numbers are optional; default is 1.
-        - Does not currently handle parentheses, charges, or hydrates.
-    """
-    if not formula:
-        raise ValueError("Empty formula string.")
-
-    pos = 0
-    element_counts_dict: Dict[str, int] = {}
-    for m in _ELNUM.finditer(formula):
-        if m.start() != pos:
-            unknown = formula[pos : m.start()]
-            raise ValueError(f"Unsupported token '{unknown}' in formula '{formula}'.")
-        elem, num = m.groups()
-        coeff = int(num) if num else 1
-        element_counts_dict[elem] = element_counts_dict.get(elem, 0) + coeff
-        pos = m.end()
-
-    if pos != len(formula):
-        unknown = formula[pos:]
-        raise ValueError(
-            f"Unsupported trailing token '{unknown}' in formula '{formula}'."
-        )
-    return element_counts_dict
-
-
-def build_formula_matrix(
-    df_molname: pd.DataFrame,
-    species_col: str = "JANAF",
-    *,
-    element_order: Optional[Sequence[str]] = None,
-    sanitize: bool = True,
-    species_names: str = "raw",  # NEW: "raw" | "clean" | "both"
-) -> Union[
-    Tuple[np.ndarray, List[str], List[str]],
-    Tuple[np.ndarray, List[str], List[str], List[str]]
-]:
-    """
-    Build a stoichiometric matrix A (elements x species) from df_molname.
-
-    Args:
-        species_names : {"raw", "clean", "both"}, default "raw"
-        Which species labels to return:
-        - "raw": return original strings (e.g., "C1N2(CNN)")
-        - "clean": return sanitized/normalized species (e.g., "C1N2")
-        - "both": return both (species_raw, species_clean)
-
-    Returns:
-        If species_names in {"raw", "clean"}: A, elements, species
-        If species_names == "both": A, elements, species_raw, species_clean
-    """
-    raw_species = df_molname[species_col].astype(str).tolist()
-
-    # Clean names for parsing / normalization (drop leading markers & trailing annotations)
-    species_clean = [sanitize_formula(s) if sanitize else s for s in raw_species]
-
-    # Parse using the clean base (never use raw for parsing)
-    parsed_list = [parse_formula_with_charge(s) for s in species_clean]
-
-    all_elements = set().union(*parsed_list) if parsed_list else set()
-    elements = sorted(all_elements) if element_order is None else list(element_order)
-
-    A = np.zeros((len(elements), len(species_clean)), dtype=np.int64)
-    elem_index = {e: i for i, e in enumerate(elements)}
-
-    for j, d in enumerate(parsed_list):
-        for e, v in d.items():
-            if e not in elem_index:
-                raise ValueError(
-                    f"Unexpected element '{e}' in species '{species_clean[j]}'."
-                )
-            A[elem_index[e], j] = int(v)
-
-    # Choose which species names to return (labels only; matrix columns already built)
-    if species_names == "raw":
-        return A, elements, raw_species
-    elif species_names == "clean":
-        return A, elements, species_clean
-    elif species_names == "both":
-        return A, elements, raw_species, species_clean
-    else:
-        raise ValueError("species_names must be 'raw', 'clean', or 'both'.")
